requirements.txt
praw
openai
pandas
streamlit
python-dotenv